+++ Wily-ios/Wily/PlaylistTableViewController.m
@@ -5,6 +5,7 @@
 
 - (void)viewDidLoad {
   [super viewDidLoad];
+  self.navigationItem.leftBarButtonItem = [[UIBarButtonItem alloc] initWithBarButtonSystemItem:UIBarButtonSystemItemOrganize target:self action:@selector(showAlert:)];
   self.navigationItem.rightBarButtonItem = self.editButtonItem;
 }
 
@@ -44,6 +45,27 @@
 
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath {
   [self.playlist setCurrentlyPlayingIndex:indexPath.row];
+}
+
+- (void)showAlert:(id)sender {
+  UIAlertController* alert = [UIAlertController alertControllerWithTitle:nil
+                                                                 message:nil
+                                                          preferredStyle:UIAlertControllerStyleActionSheet];
+  [alert addAction:
+   [UIAlertAction actionWithTitle:[self repeatAutoplayActionSheetTitle]
+                            style:UIAlertActionStyleDefault
+                          handler:^(UIAlertAction * action) {
+                            [self.playlist toggleAutoplay];
+                          }]];
+  [alert addAction:
+   [UIAlertAction actionWithTitle:@"Cancel"
+                            style:UIAlertActionStyleCancel
+                          handler:nil]];
+  [self presentViewController:alert animated:YES completion:nil];
+}
+
+- (NSString *)repeatAutoplayActionSheetTitle {
+  return self.playlist.shouldAutoplay ? @"Repeat Song Mode" : @"Playlist Mode";
 }
 
 @end