+++ Wily-ios/Wily/WilyPlayer.h
@@ -18,11 +18,11 @@ typedef enum : NSUInteger {
                               song:(NSDictionary *)song;
 
 @property (nonatomic, readonly) NSDictionary *song;
-@property (nonatomic) BOOL shouldRepeat;
 
 @property (nonatomic, weak) id<WilyPlayerDelegate> delegate;
 
 - (void)togglePlayPause;
+- (void)repeat;
 
 @property(nonatomic, readonly) WilyPlayerPlaybackState playbackState;
 @property (nonatomic, readonly) NSTimeInterval currentPlaybackTime;